requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.0
absl-py=2.3.0=pypi_0
accelerate=0.31.0=pypi_0
addict=2.4.0=pypi_0
affine=2.4.0=pypi_0
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.10.11=pypi_0
aiohttp-retry=2.8.3=pypi_0
aioice=0.10.1=pypi_0
aiortc=1.9.0=pypi_0
aiosignal=1.3.2=pypi_0
aiosqlite=0.21.0=pypi_0
albucore=0.0.24=pypi_0
albumentations=2.0.8=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.49.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.9.0=pypi_0
apscheduler=3.11.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.5=pypi_0
async-timeout=5.0.1=pypi_0
asyncua=1.1.6=pypi_0
attrs=25.3.0=pypi_0
av=12.3.0=pypi_0
babel=2.17.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.13.4=pypi_0
bitsandbytes=0.45.5=pypi_0
bleach=6.2.0=pypi_0
blinker=1.9.0=pypi_0
boto3=1.35.60=pypi_0
botocore=1.35.99=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
cachetools=5.5.2=pypi_0
certifi=2025.4.26=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
cloudpickle=3.1.1=pypi_0
colorama=0.4.6=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.2=pypi_0
copulas=0.12.3=pypi_0
cryptography=45.0.4=pypi_0
ctgan=0.11.0=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.12=pypi_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.14=pypi_0
decorator=5.2.1=pypi_0
deepecho=0.7.0=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
docker=7.1.0=pypi_0
docopt=0.6.2=pypi_0
einops=0.8.0=pypi_0
et-xmlfile=2.0.0=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.0=pypi_0
faker=37.4.2=pypi_0
fastapi=0.110.3=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.17.0=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.58.1=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.5.1=pypi_0
ftfy=6.3.1=pypi_0
gast=0.6.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
google-crc32c=1.7.1=pypi_0
google-pasta=0.2.0=pypi_0
graphviz=0.21=pypi_0
grpcio=1.71.0=pypi_0
h11=0.16.0=pypi_0
h5py=3.13.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.32.4=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.7=pypi_0
ifaddr=0.2.0=pypi_0
imageio=2.37.0=pypi_0
imbalanced-learn=0.13.0=pypi_0
inference=0.50.4=pypi_0
iniconfig=2.1.0=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.36.0=pypi_0
ipywidgets=8.1.7=pypi_0
isoduration=20.11.0=pypi_0
jax=0.6.1=pypi_0
jaxlib=0.6.1=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.10.0=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.5.1=pypi_0
json5=0.12.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.24.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.8.1=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.16.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.4.3=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.15=pypi_0
keras=3.10.0=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
libclang=18.1.1=pypi_0
libffi=3.4.4=hd77b12b_1
markdown=3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib=3.10.3=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mediapipe=0.10.14=pypi_0
mistune=3.1.3=pypi_0
ml-dtypes=0.5.1=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.4.4=pypi_0
mypy-extensions=1.1.0=pypi_0
namex=0.1.0=pypi_0
narwhals=1.48.1=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=pypi_0
notebook=7.4.3=pypi_0
notebook-shim=0.2.4=pypi_0
num2words=0.5.14=pypi_0
numpy=2.1.3=pypi_0
nvidia-ml-py=12.575.51=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.21.1=pypi_0
openai=1.86.0=pypi_0
opencv-contrib-python=4.10.0.84=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.16=h3f729d1_0
opt-einsum=3.4.0=pypi_0
optree=0.16.0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
paho-mqtt=1.6.1=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
peft=0.11.1=pypi_0
pillow=11.2.1=pypi_0
pillow-heif=0.22.0=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pypi_0
plotly=6.2.0=pypi_0
pluggy=1.6.0=pypi_0
portalocker=3.1.1=pypi_0
prometheus-client=0.22.0=pypi_0
prometheus-fastapi-instrumentator=6.0.0=pypi_0
prompt-toolkit=3.0.51=pypi_0
propcache=0.3.2=pypi_0
protobuf=4.25.8=pypi_0
psutil=7.0.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=21.0.0=pypi_0
pybase64=1.0.2=pypi_0
pycocotools=2.0.9=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.5=pypi_0
pydantic-core=2.33.2=pypi_0
pydantic-settings=2.7.1=pypi_0
pydeck=0.9.1=pypi_0
pydot=2.0.0=pypi_0
pyee=13.0.0=pypi_0
pygments=2.19.1=pypi_0
pyjwt=2.10.1=pypi_0
pylibsrtp=0.12.0=pypi_0
pylogix=1.0.5=pypi_0
pymodbus=3.8.3=pypi_0
pyopenssl=25.1.0=pypi_0
pyparsing=3.2.3=pypi_0
pyreadline3=3.5.4=pypi_0
pytest=8.4.0=pypi_0
python=3.10.16=h4607a30_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=3.3.0=pypi_0
pytz=2025.2=pypi_0
pyvips=3.0.0=pypi_0
pywavelets=1.8.0=pypi_0
pywin32=310=pypi_0
pywinpty=2.0.15=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.4.0=pypi_0
rasterio=1.4.3=pypi_0
rdt=1.17.1=pypi_0
redis=5.0.8=pypi_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-file=2.1.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rf-clip=1.1=pypi_0
rf-groundingdino=0.2.0=pypi_0
rf-segment-anything=1.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.9.4=pypi_0
roboflow=1.1.65=pypi_0
rpds-py=0.25.1=pypi_0
s3transfer=0.10.4=pypi_0
safetensors=0.5.3=pypi_0
samv2=0.0.4=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.3=pypi_0
sdmetrics=0.22.0=pypi_0
sdv=1.24.1=pypi_0
seaborn=0.13.2=pypi_0
segmentation-models-pytorch=0.5.0=pypi_0
send2trash=1.8.3=pypi_0
setuptools=78.1.1=py310haa95532_0
shapely=2.0.7=pypi_0
shellingham=1.5.4=pypi_0
simsimd=6.4.7=pypi_0
six=1.17.0=pypi_0
sklearn-compat=0.1.3=pypi_0
slack-sdk=3.33.5=pypi_0
smmap=5.0.2=pypi_0
smogn=0.1.2=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
sounddevice=0.5.2=pypi_0
soupsieve=2.7=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
starlette=0.37.2=pypi_0
streamlit=1.48.1=pypi_0
stringzilla=3.12.5=pypi_0
structlog=24.4.0=pypi_0
style=1.1.0=pypi_0
supervision=0.25.1=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.1.2=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.19.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=3.1.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.5.10=pypi_0
timm=1.0.15=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h0416ee5_0
tldextract=5.1.3=pypi_0
tokenizers=0.21.1=pypi_0
toml=0.10.2=pypi_0
tomli=2.2.1=pypi_0
torch=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tornado=6.5.1=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.51.3=pypi_0
twilio=9.3.8=pypi_0
typer=0.12.5=pypi_0
types-python-dateutil=2.9.0.20250516=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
tzlocal=5.3.1=pypi_0
ultralytics=8.3.40=pypi_0
ultralytics-thop=2.0.14=pypi_0
update=0.0.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.4.0=pypi_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
wait-for2=0.3.2=pypi_0
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310haa95532_0
widgetsnbextension=4.0.14=pypi_0
wrapt=1.17.2=pypi_0
xgboost=3.0.2=pypi_0
xz=5.6.4=h4754444_1
yapf=0.43.0=pypi_0
yarl=1.20.1=pypi_0
zlib=1.2.13=h8cc25b3_1
zxing-cpp=2.2.0=pypi_0
